+++ experimental/CHANGELOG.md
@@ -18,6 +18,8 @@ For notes on migrating to 2.x / 0.200.x see [the upgrade guide](doc/upgrade-to-2
 
 ### :bug: Bug Fixes
 
+* fix(sdk-logs): invoke all lifecycle processors during flush and shutdown [#6980](https://github.com/open-telemetry/opentelemetry-js/pull/6980) @teamleaderleo
+
 ### :books: Documentation
 
 * docs(configuration): document the `ConfigModel` suffix naming convention for exported types [#6612](https://github.com/open-telemetry/opentelemetry-js/issues/6612) @vedantchalke36